requirements.txt
Django
djangorestframework
psycopg2
gunicorn == 20.1.0
django-heroku == 0.3.1